We can flip two now!
Using the same GameObject. Let's call this a "GameController", shall we?
Now it should verify if they match or not. By the way, DO NOT FORGET:
the buttons should have in their (disabled) text the hexadecimal for the
color they represent. When flipping (Another script for the
GameController), passing 90 degrees, they show another thing (text or
color).