Animation on rotation
Not the animation at the beginning of the application, but at the
beginning of the gameplay! To be honest, This kinda developed a script
for those rotating cards that were bothering me. This remembered me how
fun it is to develop anything! Don't mind the time tho. Also updated to
Unity 5.3.4f1